+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- prometheus_exporter's LabelNames.sloApp default value now matches documented one (slo_app)
 
 ## [5.1.0] 2020-05-4
 ### Added