+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- New module `metadataClassifier` to classify event based on it's metadata. See [the docs](docs/modules/metadata_classifier.md) for more info.
 
 ## [5.1.1] 2020-05-04
 ### Changed